# Add constant for default content_tag_with_haml options