@amatsuda 2017-10-26
@oniofchaos Just manually merged this to master. Thanks!